Moved free functions to FutureUtils
made `flatMap() -> Future<U>` a free function instead of a instance of `Result`
